Port over some span manipulation from CanvasRenderer -> Span
